+++ packages/web3-shared/evm/src/index.ts
@@ -2,3 +2,4 @@ export * from './types/index.js'
 export * from './helpers/index.js'
 export * from './constants/index.js'
 export * from './libs/index.js'
+export * from './helpers/getReceiptStatus.js'
